% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sharer.R
\docType{data}
\name{sharers}
\alias{sharers}
\title{Available sharers}
\format{
An object of class \code{character} of length 25.
}
\usage{
sharers
}
\description{
Available sharers
}
\examples{
sharers
}
\keyword{datasets}
