% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sharer.R
\name{social_btns}
\alias{social_btns}
\title{Create social share buttons}
\usage{
social_btns(
  url,
  title,
  media = c("twitter", "facebook", "reddit"),
  width = "100\%",
  height = "4.25em"
)
}
\arguments{
\item{url}{url string to share.}

\item{title}{title string to share.}

\item{media}{sharers to set as widget.}

\item{width}{width of widget's wrapper.}

\item{height}{width of widget's height.}
}
\value{
An object of class \code{htmlwidget} that will intelligently print
  itself into HTML in a variety of contexts including the R console, within R
  Markdown documents, and within Shiny output bindings.
}
\description{
Create social share buttons for R Markdown or Shiny.
}
